Extract Xeger to a standalone namespace
