+++ Src/AutoFixture/Fare/Automaton.cs
@@ -34,7 +34,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace AutoFixture.DataAnnotations
+namespace Fare
 {
     /// <summary>
     /// Finite-state automaton with regular expression operations.
