+++ Src/AutoFixture/Fare/RegExpMatchingOptions.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace AutoFixture.DataAnnotations
+namespace Fare
 {
     internal static class RegExpMatchingOptions
     {
