+++ Src/AutoFixture/Fare/StateEqualityComparer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace AutoFixture.DataAnnotations
+namespace Fare
 {
     internal sealed class StateEqualityComparer : IEqualityComparer<State>
     {
