+++ Src/AutoFixture/RegularExpressionGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text.RegularExpressions;
-using AutoFixture.DataAnnotations;
 using AutoFixture.Kernel;
+using Fare;
 
 namespace AutoFixture
 {
